+++ activesupport/CHANGELOG.md
@@ -1,3 +1,6 @@
+*   Make Dependencies pass a name to NameError error.
+    *arthurnn*, *Yuki Nishijima*
+
 *   Fixed precision error in NumberHelper when using Rationals.
 
     before: